Refactors and consolidates upload-related types
