+++ client/src/components/Panels/Upload/uploadState.test.ts
@@ -1,8 +1,8 @@
 import { suppressExpectedErrorMessages } from "@tests/vitest/helpers";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
+import type { UploadCollectionConfig } from "@/composables/upload/collectionTypes";
 import type { NewUploadItem } from "@/composables/upload/uploadItemTypes";
-import type { CollectionConfig } from "@/composables/uploadQueue";
 
 import { useUploadState } from "./uploadState";
 
@@ -24,7 +24,7 @@ function makePastedItem(name = "file.txt", content = "hello world"): NewUploadIt
     };
 }
 
-const BATCH_CONFIG: CollectionConfig = {
+const BATCH_CONFIG: UploadCollectionConfig = {
     name: "My Collection",
     type: "list",
     hideSourceItems: false,
