+++ client/src/utils/upload.ts
@@ -54,9 +54,8 @@ import {
 } from "@/components/Collections/pairing";
 import type { PreparedUpload } from "@/components/Panels/Upload/types";
 import type { UploadRowModel } from "@/components/Upload/model";
-import type { SupportedCollectionType } from "@/composables/upload/collectionTypes";
+import type { SupportedCollectionType, UploadCollectionConfig } from "@/composables/upload/collectionTypes";
 import type { NewUploadItem } from "@/composables/upload/uploadItemTypes";
-import type { CollectionConfig } from "@/composables/uploadQueue";
 import { getAppRoot } from "@/onload/loadConfig";
 import { errorMessageAsString } from "@/utils/simple-error";
 import { isUrl } from "@/utils/url";
@@ -1011,7 +1010,7 @@ export async function uploadDatasets(items: ApiUploadItem[], config: UploadDatas
 /**
  * Builds a PreparedUpload object from UI upload items.
  */
-export function buildPreparedUpload(items: NewUploadItem[], collectionConfig?: CollectionConfig): PreparedUpload {
+export function buildPreparedUpload(items: NewUploadItem[], collectionConfig?: UploadCollectionConfig): PreparedUpload {
     return {
         apiItems: items.map((item) => toApiUploadItem(item)),
         collectionConfig,
